+++ src/Art.cpp
@@ -192,7 +192,7 @@ namespace Arts {
 					Art* a = new Art("compassIn part");
 
 					for (int i = 0;i < compassInSize.y;i++) {
-						if (j + i < fullCompassInArt->lines.size()) {
+						if (j + i < (int)fullCompassInArt->lines.size()) {
 							a->lines.push_back(fullCompassInArt->lines[j + i]);
 						}
 					}
@@ -287,7 +287,7 @@ namespace Arts {
     }
 
     Art* getArt(int index) {
-        if (index >= 0 && index < artList.size()) {
+        if (index >= 0 && index < (int)artList.size()) {
             return artList[index];
         } else {
             return defaultArt;
@@ -318,7 +318,7 @@ void Art::printAt(Point2 pos) {
 
 void Art::printCenter(int y, int xOffset) {
     int w = getSize().x;
-    for (int i = 0; i < lines.size(); i++) {
+    for (int i = 0; i < (int)lines.size(); i++) {
         mvprintw(y + i, (Ui::terminalSize.x / 2)+xOffset-((w - 1) / 2), lines[i].c_str());
     }
 }
